# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online hautarzt API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/doctors/register": {
            "post": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Register a new doctor",
                "description": "Register a new doctor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "phone_number",
                                    "password",
                                    "recaptcha",
                                    "accepted",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "A doctor's e-mail",
                                        "format": "string",
                                        "example": "doctor@gmail.com"
                                    },
                                    "phone_number": {
                                        "description": "A doctor's phone number",
                                        "format": "string",
                                        "example": "+3 8(032) 345-34-34"
                                    },
                                    "password": {
                                        "description": "A doctor's password",
                                        "format": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation. Must be the same as the password",
                                        "format": "string"
                                    },
                                    "recaptcha": {
                                        "description": "A recaptcha token. The action must be 'register_doctor",
                                        "format": "string"
                                    },
                                    "accepted": {
                                        "description": "Is terms and conditions accepted",
                                        "format": "boolean",
                                        "example": "1"
                                    },
                                    "board_certification": {
                                        "description": "A doctor's board certification",
                                        "format": "binary"
                                    },
                                    "medical_degree": {
                                        "description": "A doctor's medical degree",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "A new doctor has been succesfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "format": "object",
                                            "$ref": "#/components/schemas/AuthTokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "There are some validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "ValidationError",
                                    "title": "Validation error",
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "photo": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The photo field is required."
                                                    }
                                                },
                                                "email": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "password": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password field is required."
                                                    }
                                                },
                                                "accepted": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The accepted field is required."
                                                    }
                                                },
                                                "recaptcha": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The recaptcha field is required."
                                                    }
                                                }
                                            },
                                            "type": "object",
                                            "format": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Server Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthTokenResource": {
                "title": "AuthToken resource",
                "description": "Resource for an authorization token representation",
                "properties": {
                    "id": {
                        "format": "integer",
                        "example": "1"
                    },
                    "access_token": {
                        "format": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjQ3ZGY5ZDdkYmY4ZmM1Mz"
                    },
                    "expires_at": {
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    }
                },
                "type": "object"
            },
...
            "CarbonResource": {
                "title": "Carbon resource",
                "description": "Resource for datetime representation",
                "properties": {
                    "date": {
                        "format": "string",
                        "example": "2019-12-11 12:25:00.000000"
                    },
                    "timezone_type": {
                        "format": "int64",
                        "example": 3
                    },
                    "timezone": {
                        "format": "string",
                        "example": "Europe/Berlin"
                    }
                },
                "type": "object"
            },
...
        }
    }
}